Move loadPlugins to a utils file to be shared w/ dev server.
